+++ spring-oxygen-flysql/src/main/java/com/isxcode/oxygen/flysql/parse/SqlValue.java
@@ -11,6 +11,10 @@ public interface SqlValue {
         return FlysqlExecute.addSingleQuote(val);
     }
 
+    default String othersVarchar(String val) {
+        return FlysqlExecute.addSingleQuote(val);
+    }
+
     default String getBooleanValue(String val) {
         return val.toString();
     }