+++ spring-oxygen-flysql/src/test/resources/db/schema_sqlserver.sql
@@ -1,24 +1,14 @@
-DROP TABLE IF EXISTS META_DATA_T;
+IF EXISTS(Select 1 From Sysobjects Where Name='DOGS_T')
+    DROP table DOGS_T;
 
-CREATE TABLE META_DATA_T
+CREATE TABLE DOGS_T
 (
-    is_delete          INT          NOT NULL,
-    created_by         VARCHAR(100) NOT NULL,
-    created_date       TIMESTAMP    NOT NULL,
-    last_modified_by   VARCHAR(100) NOT NULL,
-    last_modified_date TIMESTAMP    NOT NULL,
-    version            INT          NOT NULL,
-    an_byte            TINYINT      NOT NULL,
-    an_int             INT          NOT NULL,
-    an_short           SMALLINT     NOT NULL,
-    an_long            BIGINT       NOT NULL,
-    an_double          DOUBLE       NOT NULL,
-    an_float           FLOAT        NOT NULL,
-    an_boolean         BOOLEAN      NOT NULL,
-    an_char            CHAR         NOT NULL,
-    an_string          VARCHAR(100) NOT NULL,
-    an_date            TIMESTAMP    NOT NULL,
-    an_local_date      TIMESTAMP    NOT NULL,
-    an_local_date_time DATETIME     NOT NULL,
-    an_big_decimal     DECIMAL      NOT NULL
+    id                    int            not null,
+    name                  varchar(100)   not null,
+    amount_double         float          not null,
+    amount_big_decimal    decimal(10, 5) not null,
+    is_alive              bit            not null,
+    birth_date            date           not null,
+    birth_local_date      date           not null,
+    birth_local_date_time datetime       not null
 );